+++ __init__.py
@@ -1,2 +1,3 @@
 import ranger.api
+
 from .oscyank import oscyank, set_oscyank